+++ club/src/main/java/com/club/club/Controller/PuntuacionController.java
@@ -2,11 +2,7 @@ package com.club.club.Controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import com.club.club.model.Puntuacion;
 import com.club.club.repository.CompeticionRepository;
@@ -55,7 +51,7 @@ public class PuntuacionController {
 
     // Editar puntuación
     @GetMapping("/editar/{id}")
-    public String editarPuntuacion(@PathVariable Integer id, Model model) {
+    public String editarPuntuacion(@PathVariable Long id, Model model) {
         Puntuacion puntuacion = puntuacionRepository.findById(id).orElseThrow();
         model.addAttribute("puntuacion", puntuacion);
         model.addAttribute("gimnastas", gimnastaRepository.findAll());
@@ -65,7 +61,7 @@ public class PuntuacionController {
 
     // Borrar puntuación
     @GetMapping("/borrar/{id}")
-    public String borrarPuntuacion(@PathVariable Integer id) {
+    public String borrarPuntuacion(@PathVariable Long id) {
         puntuacionRepository.deleteById(id);
         return "redirect:/puntuaciones";
     }
